+++ web-builder/changelog.md
@@ -6,9 +6,10 @@
 > Commit 信息格式遵循[约定式提交（Conventional Commits）](http://conventionalcommits.org)标准。
 > 本变更日志遵循 [Keep a Changelog](http://keepachangelog.com/) 格式编写。
 
-## 0.2.1
+## 1.0.0
 <!-- Unreleased -->
 ### 新增
+- 增加了`init` `clone` `initConfig` `build` `verifyConfig` `pack` `help` 命令
 - 增加了`config.yml`的覆盖警告
 ### 变更
 - `configBuilder.ts`改为从release的assets中获取
@@ -17,7 +18,7 @@
 <!--/ Unreleased -->
 
 ## 0.2.0
-初步编写中。完成了`clone`，`initConfig`的函数体。
+初步编写中。完成了`clone`，`initConfig`的函数体
 ### 新增
 - 新增了`clone`函数
 - 新增了`initConfig`函数